add threading for non-rendering work
This commit implements a framework for communicating and scheduling
blocking/io operations on threads other than the rendering thread. This
is achieved with a trait that defines a model and a message type, and
a protocol for generating messages, running blocking operations in
response to them, and finally updating the model. Messages are sent
from the rendering thread, which also handles input through an
immediate mode gui, to an async function (event handler) running on a
tokio executor. The async function gets a model wrapped in a mutex,
giving it more granularity over blocking the renderer than blocking
for whole event handler execution.