refactor to use glium
Macroquad's abstractions were getting in the way of handling input. The
exact issue is that I needed to get events to recognize a mouse drag on
the background of the window, that is not on egui. After searching
through egui's and macroquads' docs I couldn't find an easy way to do
it, so the hard was to use glutin+glium which gives you events directly.
This also removes some async that was only needed for macroquad.

One problem with glium the API is too safe and egui_glium doesn't work
with raw gl IDs, and require that I fake clone textures for egui_glium
to work with later. run_app also got significantly uglier but that's to
be expected with a low level API. Oh, and having to pass around &screen
all the time is a bit annoying...